index: integrate sparse and full modes
