+++ CHANGELOG.md
@@ -7,8 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [unreleased]
 
 ### Added
+- Add (almost) full index. The database is still pruned,
+  i.e. empty files and inter-repo duplicates are removed).
 
 ### Changed
+- Change database schema: columns renamed and reordered.
 
 ### Deprecated
